Moved BinaryPriority BinaryPriorityCostCalculator and InvariantHueristicQueue into new headers
